// src/common/env.types.ts
export enum EnvKey {
  PORT = 'PORT',
  NODE_ENV = 'NODE_ENV',
  MONGODB_URI = 'MONGODB_URI',
  MONGODB_CORE_DBNAME = 'MONGODB_CORE_DBNAME',
  LOGGER_LEVEL = 'LOGGER_LEVEL',
  OTEL_EXPORTER_OTLP_TRACES_ENDPOINT = 'OTEL_EXPORTER_OTLP_TRACES_ENDPOINT',
  PROMETHEUS_PORT = 'PROMETHEUS_PORT',
  APPLICATION_NAME = 'APPLICATION_NAME',
  APPLICATION_VERSION = 'APPLICATION_VERSION',
  SERVER_CONTEXT_PATH = 'SERVER_CONTEXT_PATH',
  IDM_AUTH_SERVICE_URL = 'IDM_AUTH_SERVICE_URL',
  IDM_AUTH_APPLICATION_REALM_PUBLIC_UUID = 'IDM_AUTH_APPLICATION_REALM_PUBLIC_UUID',
}